+++ app/controllers/HomeController.scala
@@ -1,0 +1,14 @@
+package controllers
+
+import javax.inject._
+import play.api.mvc._
+
+@Singleton
+class HomeController @Inject()(val controllerComponents: ControllerComponents)
+  extends BaseController {
+
+  def index(): Action[AnyContent] = Action { implicit req =>
+    Redirect(routes.TodoController.index())
+  }
+
+}
