+++ app/views/todo/Index.scala.html
@@ -1,0 +1,41 @@
+@import helper._
+@import requests.TodoForm
+
+@(vv: model.ViewValueCommon, todos: Seq[model.ViewValueTodo], todoForm: Form[TodoForm])(implicit req: RequestHeader, message: Messages)
+
+@common.Default(vv){
+ <div class="uk-text-right uk-margin-bottom">
+  <a href="/todo/create" >+新規作成</a>
+ </div>
+ @for(todo <- todos) {
+  <div class="card bg-white uk-flex uk-flex-between uk-flex-middle uk-margin-bottom">
+   <div class="card-left">
+    <div class="uk-margin-small-bottom">
+     <span class="uk-margin-small-bottom">@todo.status.name</span>
+     <span style="margin: 0 4px;">|</span>
+     <span class="category" style="background-color: @todo.categoryColor.color;">@todo.categoryName</span>
+    </div>
+    <h3 class="card__title uk-margin-remove-top uk-margin-remove-bottom">@todo.title</h3>
+    <p class="uk-margin-remove">@todo.body</p>
+   </div>
+   <div class="card-right uk-flex uk-flex-middle uk-flex-between" style="min-width: 68px">
+    <a href="@routes.TodoController.update(todo.id)" uk-icon="icon: file-edit; ratio: 1.2" class="a-hover-none uk-padding-small"></a>
+    @form(routes.TodoController.removeAction(todo.id)) {
+     @CSRF.formField
+     <input type="hidden" name="_method" value="DELETE">
+     <button type="submit" onclick="confirmFromSubmit(event, '@todo.title')" uk-icon="icon: trash; ratio: 1.2" class="uk-padding-small"></button>
+    }
+   </div>
+  </div>
+ }
+}
+
+<script>
+    function confirmFormSubmit(event, title) {
+      event.preventDefault();
+
+      if (confirm(`${title}を削除しますか`)) {
+        document.getElementById("myForm").submit();
+      }
+    }
+</script>
