+++ conf/routes
@@ -1,9 +1,11 @@
 
 GET / controllers.HomeController.index
+GET /todo controllers.TodoController.index
 GET /todo/create controllers.TodoController.create
 POST /todo/create controllers.TodoController.createAction
 GET /todo/update/:id controllers.TodoController.update(id: Long)
 POST /todo/update/:id controllers.TodoController.updateAction(id: Long)
+POST /todo/delete/:id controllers.TodoController.removeAction(id: Long)
 
 # Map static resources from the /public folder to the /assets URL path
 GET /assets/*file controllers.Assets.versioned(path="/public", file: Asset)
